pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "opencv-python>=4.10.0.84",
    "pillow==11.0.0",
    "requests==2.32.3",
]
